Additional change for different systems
The KeepOpen bit is working; unfortunately, whether the input stream is seekable depends on the OS/.Net version. Therefore, I have changed the code to cover all versions. The input stream is copied into a memory stream, and the RawBody and Body are created from the memory stream. This works on all systems.